+++ src/main/java/com/global/api/network/entities/nts/INtsRequestMessage.java
@@ -148,14 +148,14 @@ public interface INtsRequestMessage {
 
 
         // Prior Message Response Time
-        NtsUtils.log("Prior Message Response Time", String.valueOf(ntsRequestMessageHeader.getPriorMessageResponseTime()));
-        headerRequest.addRange(StringUtils.padLeft(ntsRequestMessageHeader.getPriorMessageResponseTime(), 3, '0'), 3);
+        NtsUtils.log("Prior Message Response Time", String.valueOf(ntsRequestMessageHeader.getPriorMessageInfo().getPriorMessageResponseTime()));
+        headerRequest.addRange(StringUtils.padLeft(ntsRequestMessageHeader.getPriorMessageInfo().getPriorMessageResponseTime(), 3, '0'), 3);
         // Prior Message Connect Time
-        NtsUtils.log("Prior Message Connect Time", String.valueOf(ntsRequestMessageHeader.getPriorMessageConnectTime()));
-        headerRequest.addRange(ntsRequestMessageHeader.getPriorMessageConnectTime(), 3);
+        NtsUtils.log("Prior Message Connect Time", String.valueOf(ntsRequestMessageHeader.getPriorMessageInfo().getPriorMessageConnectTime()));
+        headerRequest.addRange(ntsRequestMessageHeader.getPriorMessageInfo().getPriorMessageConnectTime(), 3);
         // Prior Message Code
-        NtsUtils.log("Prior Message Code", String.valueOf(ntsRequestMessageHeader.getPriorMessageCode()));
-        headerRequest.addRange(ntsRequestMessageHeader.getPriorMessageCode(), 2);
+        NtsUtils.log("Prior Message Code", String.valueOf(ntsRequestMessageHeader.getPriorMessageInfo().getPriorMessageCode()));
+        headerRequest.addRange(ntsRequestMessageHeader.getPriorMessageInfo().getPriorMessageCode(), 2);
 
         NtsUtils.log("Request header :", headerRequest.getMessageRequest().toString());