+++ src/test/java/com/global/api/tests/network/nts/NTSPOSConfigurationMessageTest.java
@@ -7,6 +7,7 @@ import com.global.api.entities.enums.*;
 import com.global.api.entities.exceptions.ApiException;
 import com.global.api.network.entities.nts.NtsRequestMessageHeader;
 import com.global.api.network.entities.nts.POSSiteConfigurationData;
+import com.global.api.network.entities.nts.PriorMessageInfo;
 import com.global.api.network.enums.CardDataInputCapability;
 import com.global.api.network.enums.CardHolderAuthenticationCapability;
 import com.global.api.network.enums.TerminalOutputCapability;
@@ -23,6 +24,7 @@ import static org.junit.Assert.assertNotNull;
 
 public class NTSPOSConfigurationMessageTest {
     private NtsRequestMessageHeader ntsRequestMessageHeader;
+    private PriorMessageInfo priorMessageInfo;
     // gateway config
     NetworkGatewayConfig config;
     public NTSPOSConfigurationMessageTest() throws ApiException {
@@ -42,9 +44,12 @@ public class NTSPOSConfigurationMessageTest {
             ntsRequestMessageHeader.setTerminalDestinationTag("478");
             ntsRequestMessageHeader.setPinIndicator(PinIndicator.WithPin);
             ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.PosSiteConfiguration);
-            ntsRequestMessageHeader.setPriorMessageResponseTime(999);
-            ntsRequestMessageHeader.setPriorMessageConnectTime(999);
-            ntsRequestMessageHeader.setPriorMessageCode("08");
+            priorMessageInfo=new PriorMessageInfo();
+            priorMessageInfo.setPriorMessageResponseTime(999);
+            priorMessageInfo.setPriorMessageConnectTime(999);
+            priorMessageInfo.setPriorMessageCode("08");
+
+            ntsRequestMessageHeader.setPriorMessageInfo(priorMessageInfo);
 
             // data code values
             // acceptorConfig.setCardDataInputCapability(CardDataInputCapability.ContactlessEmv_ContactlessMsd_KeyEntry);