+++ src/test/java/com/global/api/tests/network/nts/NtsEMVPDLTest.java
@@ -10,6 +10,7 @@ import com.global.api.network.entities.nts.NtsEMVPDLResponse;
 import com.global.api.network.entities.nts.NtsRequestMessageHeader;
 import com.global.api.network.entities.NtsPDLData;
 import com.global.api.network.entities.emvpdl.*;
+import com.global.api.network.entities.nts.PriorMessageInfo;
 import com.global.api.network.enums.CardDataInputCapability;
 import com.global.api.network.enums.CardHolderAuthenticationCapability;
 import com.global.api.network.enums.TerminalOutputCapability;
@@ -35,6 +36,7 @@ public class NtsEMVPDLTest {
     NetworkGatewayConfig config;
     NtsRequestMessageHeader ntsRequestMessageHeader; //Main Request header class
     private EMVPDLTable10V2 table10V2;
+    private PriorMessageInfo priorMessageInfo;
 
     public NtsEMVPDLTest() throws ConfigurationException {
         Address address = new Address();
@@ -53,10 +55,12 @@ public class NtsEMVPDLTest {
         ntsRequestMessageHeader.setPinIndicator(PinIndicator.NotPromptedPin);
         ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.EmvParameterDataLoad);
 
-        ntsRequestMessageHeader.setPriorMessageResponseTime(1);
-        ntsRequestMessageHeader.setPriorMessageConnectTime(999);
-        ntsRequestMessageHeader.setPriorMessageCode("01");
+        priorMessageInfo=new PriorMessageInfo();
+        priorMessageInfo.setPriorMessageResponseTime(1);
+        priorMessageInfo.setPriorMessageConnectTime(999);
+        priorMessageInfo.setPriorMessageCode("01");
 
+        ntsRequestMessageHeader.setPriorMessageInfo(priorMessageInfo);
 
         acceptorConfig.setTerminalOutputCapability(TerminalOutputCapability.None);
         acceptorConfig.setCardDataInputCapability(CardDataInputCapability.ContactlessEmv_ContactlessMsd_KeyEntry);