+++ src/test/java/com/global/api/tests/network/nts/NtsFleetTest.java
@@ -11,6 +11,7 @@ import com.global.api.network.entities.FleetData;
 import com.global.api.network.entities.NtsProductData;
 import com.global.api.network.entities.nts.NtsRequestMessageHeader;
 import com.global.api.network.entities.NtsTag16;
+import com.global.api.network.entities.nts.PriorMessageInfo;
 import com.global.api.network.enums.*;
 import com.global.api.paymentMethods.CreditCardData;
 import com.global.api.paymentMethods.CreditTrackData;
@@ -40,6 +41,7 @@ public class NtsFleetTest {
     private NetworkGatewayConfig config;
     private NtsProductData productData;
     private FleetData fleetData;
+    private PriorMessageInfo priorMessageInfo;
 
     public NtsFleetTest() throws ApiException {
 
@@ -82,9 +84,13 @@ public class NtsFleetTest {
         ntsRequestMessageHeader.setPinIndicator(PinIndicator.NotPromptedPin);
         ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.AuthorizationOrBalanceInquiry);
 
-        ntsRequestMessageHeader.setPriorMessageResponseTime(1);
-        ntsRequestMessageHeader.setPriorMessageConnectTime(999);
-        ntsRequestMessageHeader.setPriorMessageCode("01");
+
+        priorMessageInfo=new PriorMessageInfo();
+        priorMessageInfo.setPriorMessageResponseTime(1);
+        priorMessageInfo.setPriorMessageConnectTime(999);
+        priorMessageInfo.setPriorMessageCode("01");
+
+        ntsRequestMessageHeader.setPriorMessageInfo(priorMessageInfo);
 
         fleetData = new FleetData();
         fleetData.setOdometerReading("1234567");