+++ src/test/java/com/global/api/tests/network/nts/NtsMailTest.java
@@ -9,6 +9,7 @@ import com.global.api.entities.exceptions.GatewayException;
 import com.global.api.network.entities.nts.NtsMailData;
 import com.global.api.network.entities.nts.NtsMailResponse;
 import com.global.api.network.entities.nts.NtsRequestMessageHeader;
+import com.global.api.network.entities.nts.PriorMessageInfo;
 import com.global.api.network.enums.CardDataInputCapability;
 import com.global.api.network.enums.CardHolderAuthenticationCapability;
 import com.global.api.network.enums.TerminalOutputCapability;
@@ -27,6 +28,7 @@ import static org.junit.Assert.assertNotNull;
 
 public class NtsMailTest {
     private NtsRequestMessageHeader ntsRequestMessageHeader;
+    private PriorMessageInfo priorMessageInfo;
     // gateway config
     NetworkGatewayConfig config;
     public NtsMailTest() throws ApiException {
@@ -46,9 +48,12 @@ public class NtsMailTest {
             ntsRequestMessageHeader.setTerminalDestinationTag("478");
             ntsRequestMessageHeader.setPinIndicator(PinIndicator.WithPin);
             ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.PinDebit);
-            ntsRequestMessageHeader.setPriorMessageResponseTime(999);
-            ntsRequestMessageHeader.setPriorMessageConnectTime(999);
-            ntsRequestMessageHeader.setPriorMessageCode("08");
+            priorMessageInfo=new PriorMessageInfo();
+            priorMessageInfo.setPriorMessageResponseTime(999);
+            priorMessageInfo.setPriorMessageConnectTime(999);
+            priorMessageInfo.setPriorMessageCode("08");
+
+            ntsRequestMessageHeader.setPriorMessageInfo(priorMessageInfo);
 
             // data code values
             // acceptorConfig.setCardDataInputCapability(CardDataInputCapability.ContactlessEmv_ContactlessMsd_KeyEntry);