+++ src/test/java/com/global/api/tests/network/nts/NtsPDLTest.java
@@ -9,6 +9,7 @@ import com.global.api.entities.exceptions.ConfigurationException;
 import com.global.api.network.entities.NtsPDLData;
 import com.global.api.network.entities.nts.NtsPDLResponseData;
 import com.global.api.network.entities.nts.NtsRequestMessageHeader;
+import com.global.api.network.entities.nts.PriorMessageInfo;
 import com.global.api.network.enums.CardDataInputCapability;
 import com.global.api.network.enums.CardHolderAuthenticationCapability;
 import com.global.api.network.enums.TerminalOutputCapability;
@@ -27,6 +28,7 @@ public class NtsPDLTest {
     // gateway config
     NetworkGatewayConfig config;
     NtsRequestMessageHeader ntsRequestMessageHeader; //Main Request header class
+    private PriorMessageInfo priorMessageInfo;
 
    public NtsPDLTest() throws ConfigurationException {
        Address address = new Address();
@@ -44,9 +46,12 @@ public class NtsPDLTest {
        ntsRequestMessageHeader.setTerminalDestinationTag("510");
        ntsRequestMessageHeader.setPinIndicator(PinIndicator.NotPromptedPin);
        ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.ParameterDataLoad);
-       ntsRequestMessageHeader.setPriorMessageResponseTime(1);
-       ntsRequestMessageHeader.setPriorMessageConnectTime(999);
-       ntsRequestMessageHeader.setPriorMessageCode("01");
+        priorMessageInfo=new PriorMessageInfo();
+        priorMessageInfo.setPriorMessageResponseTime(1);
+        priorMessageInfo.setPriorMessageConnectTime(999);
+        priorMessageInfo.setPriorMessageCode("01");
+
+        ntsRequestMessageHeader.setPriorMessageInfo(priorMessageInfo);
 
 
        acceptorConfig.setTerminalOutputCapability(TerminalOutputCapability.None);