+++ src/test/java/com/global/api/tests/network/nts/NtsRequestToBalanceTest.java
@@ -8,6 +8,7 @@ import com.global.api.entities.exceptions.ApiException;
 import com.global.api.entities.exceptions.ConfigurationException;
 import com.global.api.network.entities.nts.NtsRequestMessageHeader;
 import com.global.api.network.entities.nts.NtsRequestToBalanceData;
+import com.global.api.network.entities.nts.PriorMessageInfo;
 import com.global.api.network.enums.CardDataInputCapability;
 import com.global.api.network.enums.CardHolderAuthenticationCapability;
 import com.global.api.network.enums.TerminalOutputCapability;
@@ -27,6 +28,7 @@ public class NtsRequestToBalanceTest {
     // gateway config
     NetworkGatewayConfig config;
     private NtsRequestMessageHeader ntsRequestMessageHeader;
+    private PriorMessageInfo priorMessageInfo;
 
     public NtsRequestToBalanceTest() throws ConfigurationException {
         Address address = new Address();
@@ -44,9 +46,12 @@ public class NtsRequestToBalanceTest {
         ntsRequestMessageHeader.setTerminalDestinationTag("478");
         ntsRequestMessageHeader.setPinIndicator(PinIndicator.WithPin);
         ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.RequestToBalacnce);
-        ntsRequestMessageHeader.setPriorMessageResponseTime(999);
-        ntsRequestMessageHeader.setPriorMessageConnectTime(999);
-        ntsRequestMessageHeader.setPriorMessageCode("08");
+        priorMessageInfo=new PriorMessageInfo();
+        priorMessageInfo.setPriorMessageResponseTime(999);
+        priorMessageInfo.setPriorMessageConnectTime(999);
+        priorMessageInfo.setPriorMessageCode("08");
+
+        ntsRequestMessageHeader.setPriorMessageInfo(priorMessageInfo);
 
         // data code values
         // acceptorConfig.setCardDataInputCapability(CardDataInputCapability.ContactlessEmv_ContactlessMsd_KeyEntry);