+++ src/test/java/com/global/api/tests/network/nts/NtsVisaReadyLinkTest.java
@@ -7,6 +7,7 @@ import com.global.api.entities.enums.*;
 import com.global.api.entities.exceptions.ApiException;
 import com.global.api.entities.exceptions.ConfigurationException;
 import com.global.api.network.entities.nts.NtsRequestMessageHeader;
+import com.global.api.network.entities.nts.PriorMessageInfo;
 import com.global.api.network.enums.CardDataInputCapability;
 import com.global.api.network.enums.CardHolderAuthenticationCapability;
 import com.global.api.network.enums.OperatingEnvironment;
@@ -32,6 +33,7 @@ public class NtsVisaReadyLinkTest {
     private CreditTrackData track;
     private CreditCardData card;
     private NtsRequestMessageHeader ntsRequestMessageHeader;
+    private PriorMessageInfo priorMessageInfo;
 
     public NtsVisaReadyLinkTest() throws ConfigurationException {
         Address address = new Address();
@@ -50,9 +52,12 @@ public class NtsVisaReadyLinkTest {
         ntsRequestMessageHeader.setPinIndicator(PinIndicator.NotPromptedPin);
         ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.PinDebit);
 
-        ntsRequestMessageHeader.setPriorMessageResponseTime(999);
-        ntsRequestMessageHeader.setPriorMessageConnectTime(999);
-        ntsRequestMessageHeader.setPriorMessageCode("08");
+        priorMessageInfo=new PriorMessageInfo();
+        priorMessageInfo.setPriorMessageResponseTime(999);
+        priorMessageInfo.setPriorMessageConnectTime(999);
+        priorMessageInfo.setPriorMessageCode("08");
+
+        ntsRequestMessageHeader.setPriorMessageInfo(priorMessageInfo);
 
         // data code values
         // acceptorConfig.setCardDataInputCapability(CardDataInputCapability.ContactlessEmv_ContactlessMsd_KeyEntry);