+++ src/test/java/com/global/api/tests/network/nts/certification/StandInProcessing.java
@@ -10,6 +10,7 @@ import com.global.api.network.entities.FleetData;
 import com.global.api.network.entities.NtsProductData;
 import com.global.api.network.entities.NtsTag16;
 import com.global.api.network.entities.nts.NtsRequestMessageHeader;
+import com.global.api.network.entities.nts.PriorMessageInfo;
 import com.global.api.network.enums.*;
 import com.global.api.network.enums.nts.AvailableProductsCapability;
 import com.global.api.paymentMethods.CreditCardData;
@@ -38,6 +39,7 @@ public class StandInProcessing {
     private NetworkGatewayConfig config;
     private FleetData fleetData;
     private NtsProductData productData;
+    private PriorMessageInfo priorMessageInfo;
 
 
     public StandInProcessing() throws ConfigurationException {
@@ -80,9 +82,13 @@ public class StandInProcessing {
         ntsRequestMessageHeader.setTerminalDestinationTag("510");
         ntsRequestMessageHeader.setPinIndicator(PinIndicator.NotPromptedPin);
         ntsRequestMessageHeader.setNtsMessageCode(NtsMessageCode.AuthorizationOrBalanceInquiry);
-        ntsRequestMessageHeader.setPriorMessageResponseTime(1);
-        ntsRequestMessageHeader.setPriorMessageConnectTime(999);
-        ntsRequestMessageHeader.setPriorMessageCode("01");
+
+        priorMessageInfo=new PriorMessageInfo();
+        priorMessageInfo.setPriorMessageResponseTime(1);
+        priorMessageInfo.setPriorMessageConnectTime(999);
+        priorMessageInfo.setPriorMessageCode("01");
+
+        ntsRequestMessageHeader.setPriorMessageInfo(priorMessageInfo);
 
         fleetData = new FleetData();
         fleetData.setPurchaseDeviceSequenceNumber("12345");